aula4.3

+++ byte-bank/src/main/java/br/com/alura/bytebank/domain/SaldoResponse.java
@@ -1,6 +1,8 @@
 package br.com.alura.bytebank.domain;
 
+import java.math.BigDecimal;
+
 public record SaldoResponse(
-        Double saldo
+        BigDecimal saldo
 ) {
 }
